bottomNavigation 홈스크린 위에 로고도 넣었어요 버튼은 추가 아직 안함요
