+++ my-app/src/nav/BottomTabScreen.tsx
@@ -1,0 +1,50 @@
+import React from "react";
+import { createNativeStackNavigator } from "@react-navigation/native-stack";
+import { NavigationContainer } from "@react-navigation/native";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import HomeScreen from "../screens/HomeScreen";
+import DetailScreen from "../screens/DetailScreen";
+import ProfileScreen from "../screens/ProfileScreen";
+import SearchScreen from "../screens/SearchScreen";
+import ReelsScreen from "../screens/ReelsScreen";
+import Ionic from "react-native-vector-icons/Ionicons";
+
+const BottomTabScreen = () => {
+
+  const Tab = createBottomTabNavigator();
+  const Stack = createNativeStackNavigator();
+
+  return (
+    <Tab.Navigator screenOptions={({ route }) => ({
+      tabBarShowLabel: false,
+      headerShown: false,
+      tabBarStyle: {
+      height: 50,
+      },
+
+      tabBarIcon: ({ focused, color, size }) => {
+        let iconName;
+        if (route.name === 'Instagram') {
+          iconName = focused ? 'home-sharp' : "home-outline";
+        } else if (route.name === 'Search') {
+          iconName = focused ? 'search' : "ios-search-outline";
+        } else if (route.name === 'Reels') {
+          iconName = focused ? 'caret-forward-circle' : "caret-forward-circle-outline";
+        } else if (route.name === 'Profile') {
+          iconName = focused ? 'ios-person-circle' : "ios-person-circle-outline";
+        } else if (route.name === 'Detail') {
+          iconName = focused ? 'ios-heart' : "ios-heart-outline";
+        } 
+        return <Ionic name={iconName} size={size} color={color} />;
+      },
+    })}>
+      <Tab.Screen name="Instagram" component={HomeScreen}  />
+      <Tab.Screen name="Search" component={SearchScreen} />
+      <Tab.Screen name="Reels" component={ReelsScreen} />
+      <Tab.Screen name="Detail" component={DetailScreen} />
+      <Tab.Screen name="Profile" component={ProfileScreen} />
+    </Tab.Navigator>
+  )
+}
+
+export default BottomTabScreen;
