+++ my-app/src/screens/HomeScreen.tsx
@@ -1,21 +1,29 @@
+
 import React from "react";
 import { StyleSheet, Text, View } from "react-native";
+import FontAwesome from "react-native-vector-icons/FontAwesome";
+import Feather from "react-native-vector-icons/Feather";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 function HomeScreen() {
   return (
-    <View style={styles.container}>
-      <Text>Home Screen</Text>
-    </View>
+    <SafeAreaView>
+      <View style={{
+        justifyContent: "space-between",
+        flexDirection: "row",
+        paddingHorizontal: 15,
+        alignItems: "center",
+      }}>
+        <FontAwesome name="plus-square-o" style={{fontSize: 24}}/>
+        <Text style={{fontFamily: "Lobster-Regular", fontSize: 25, fontWeight: '500'}}>
+          Instagram
+        </Text>
+        <Feather name="navigation" style={{fontSize: 24}}/>
+      </View>
+    </SafeAreaView>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+
 
 export default HomeScreen;
